# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vault API
  version: "1.0.0"
  description: >
    REST surface of Vault's services: id (OIDC provider), market (listings +
    orders), pay (escrow ledger + wallet), and assist (AI suggestions + trust).
    In the running stack the browser reaches id through a same-origin `/idp/*`
    proxy; the other services are called server-side by the Next.js app. Money is
    int64 minor units (JPY). Bearer tokens are RS256 access tokens issued by id
    and verified against its JWKS.
  license:
    name: MIT
servers:
  - url: http://localhost:8081
    description: id (OIDC provider)
  - url: http://localhost:8082
    description: market (listings + orders)
  - url: http://localhost:8083
    description: pay (escrow ledger + wallet)
  - url: http://localhost:8084
    description: assist (AI suggestions + trust)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token from id's /token endpoint.
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Shared secret for service-to-service escrow calls (market → pay).
  schemas:
...
    Listing:
      type: object
      properties:
        id: { type: string, format: uuid }
        seller_id: { type: string, format: uuid }
        category_id: { type: [integer, "null"] }
        title: { type: string }
        description: { type: string }
        price_minor: { type: integer, format: int64 }
        currency: { type: string, example: JPY }
        status: { type: string, enum: [draft, active, reserved, sold, withdrawn] }
        image_url: { type: string }
        created_at: { type: string, format: date-time }
...
    Suggestion:
      type: object
      properties:
        suggestion_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        category_slug: { type: string }
        price_low: { type: [integer, "null"], format: int64 }
        price_high: { type: [integer, "null"], format: int64 }
        model: { type: string, description: "claude-opus-4-8 or 'heuristic'" }
    RiskScore:
      type: object
      properties:
        id: { type: integer }
        subject_type: { type: string, enum: [listing, order] }
        subject_id: { type: string }
        score: { type: number, format: double }
        reasons: { type: array, items: { type: string } }
        status: { type: string, enum: [queued, approved, rejected] }
        created_at: { type: string, format: date-time }
paths:
...
  /listings:
    get:
      servers: [{ url: "http://localhost:8082" }]
      summary: List active listings (FTS + keyset pagination)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: integer } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 24, maximum: 100 } }
      responses:
        "200":
          description: Page of listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Listing" } }
                  next_cursor: { type: string }
    post:
      servers: [{ url: "http://localhost:8082" }]
      summary: Create a listing (also writes a listing.created outbox event)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, price_minor]
              properties:
                title: { type: string, maxLength: 120 }
                description: { type: string }
                price_minor: { type: integer, format: int64, minimum: 1 }
                category_id: { type: [integer, "null"] }
                image_url: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Listing" } } } }
        "401": { description: Missing/invalid token }
  /listings/{id}:
    get:
      servers: [{ url: "http://localhost:8082" }]
      summary: Get a listing
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Listing" } } } }
        "404": { description: Not found }
...
  /suggest:
    post:
      servers: [{ url: "http://localhost:8084" }]
      summary: Suggest a listing (VLM when configured, else heuristic) + price band
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image_url: { type: string }
                title_hint: { type: string, maxLength: 120 }
      responses:
        "200": { description: Suggestion, content: { application/json: { schema: { $ref: "#/components/schemas/Suggestion" } } } }
        "400": { description: Provide an image_url or a title_hint }
  /suggestions/{id}/outcome:
    post:
      servers: [{ url: "http://localhost:8084" }]
      summary: Record which suggested fields the human kept (acceptance metric)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accepted_fields: { type: array, items: { type: string, enum: [title, description, category, price] } }
                listing_id: { type: [string, "null"], format: uuid }
      responses:
        "200": { description: Recorded }
...
  /admin/trust:
    get:
      servers: [{ url: "http://localhost:8084" }]
      summary: The trust review queue (admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Risk scores, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/RiskScore" } } } } }
  /admin/trust/{id}/resolve:
    post:
      servers: [{ url: "http://localhost:8084" }]
      summary: Approve or reject a queued risk score (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [action], properties: { action: { type: string, enum: [approve, reject] } } }
      responses:
        "200": { description: Resolved }
        "409": { description: Not queued }